Moving csv function to utils
